+++ includes/constants.h
@@ -3,27 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   constants.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/03 08:19:36 by emadriga          #+#    #+#             */
-/*   Updated: 2021/12/19 15:47:14 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/01/15 10:14:56 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef CONSTANTS_H
 # define CONSTANTS_H
-# define MS_PROMPT "Minishell> "
-# define MS_PROMPT_SIG_INT "\033[KMinishell>   \n"
-# define MSG_NEW_LINE "%s\n"
-# define MSG_EXIT_MINISHELL "\033[AMinishell> exit\n"
 # define LIT_OLDPWD "OLDPWD"
-# define LIT_OLDPWD_LIKE "OLDPWD="
 # define LIT_PWD_LIKE "PWD="
-# define LIT_HOME_LIKE "HOME="
-# define LIT_SHLVL_LIKE "SHLVL="
-# define LIT_SHLVL_LIKE_ONE "SHLVL=1"
-# define LIT_EXEC_MINISHELL "./minishell"
-# define LIT_LAST_CMD "_="
+# define NONE 0
 
 enum e_free_status
 {
@@ -33,8 +24,7 @@ enum e_free_status
 
 enum e_quotes_status
 {
-	NONE,
-	SINGLE,
+	SINGLE = 1,
 	DOUBLE
 };
 
@@ -44,4 +34,21 @@ enum e_bool
 	TRUE
 };
 
+enum e_pipe_fd
+{
+	READ_END,
+	WRITE_END
+};
+
+enum e_token
+{
+	APPENDS_OUTPUT_REDIRECT = 1,
+	BUILTIN,
+	COMMAND,
+	INPUT_REDIRECT,
+	HEREDOC,
+	OUPUT_REDIRECT,
+	PIPE,
+	TEXT
+};
 #endif