+++ src/minishell1.c
@@ -1,27 +1,76 @@
 #include "minishell.h"
 
-static void	processline(void)
-{
-	char	*str_got;
-	char	**param_list;
+#define CARRIAGE_RETURN "\033[AMinishell> "
+#define MSG_EXIT_MINISHELL "exit\n"
+#define MS_PROMPT "Minishell> "
 
-	str_got = readline(MS_PROMPT);
-	if (str_got == NULL)
+/**
+ * * Disables CTRL hotkey(+c) from printing ^C
+*/
+static void	disable_ctrl_c_hotkey(void)
+{
+	struct termios	termios;
+	int				rc;
+
+	rc = tcgetattr(0, &termios);
+	if (rc)
 	{
-		printf(MSG_EXIT_MINISHELL);
-		exit(0);
+		perror("tcgetattr");
+		exit(1);
 	}
-	if (*str_got != '\0')
+	termios.c_lflag &= ~ECHOCTL;
+	rc = tcsetattr(0, 0, &termios);
+	if (rc)
 	{
-		add_history(str_got);
-		if (!qm_error_detector(str_got) && has_token(str_got))
+		perror("tcsetattr");
+		exit(1);
+	}
+}
+
+/**
+ * * Exits from MS (EOF received on readline)
+ * @param ignored_env	boolean to emulate bash CTRL hotkey(+d)(exit) depending
+ *  on env was ignored at start of the program
+*/
+static void	ms_eof_exit(int ignored_env)
+{
+	if (ignored_env == FALSE)
+		printf(CARRIAGE_RETURN);
+	lst_str_free(&g_var.env);
+	printf(MSG_EXIT_MINISHELL);
+	exit(0);
+}
+
+/**
+ * * Process every input line sended on STDIN_FILENO
+ * @param ignored_env	boolean to emulate bash CTRL hotkey(+d)(exit) depending
+ *  on env was ignored at start of the program
+*/
+static void	processline(int ignored_env)
+{
+	char	*line_read;
+	char	**token_list;
+	t_str	*heredoc_list;
+
+	heredoc_list = NULL;
+	line_read = readline(MS_PROMPT);
+	if (line_read == NULL)
+		ms_eof_exit(ignored_env);
+	line_read = close_quotes_pipedfork(line_read);
+	if (*line_read != '\0')
+	{
+		add_history(line_read);
+		if (!qm_error_detector(line_read) && has_token(line_read))
 		{
-			param_list = get_tokens(str_got);
-			if (!has_pipe_redir_open(param_list))
-				new_redirections(param_list, &g_var.env);
+			token_list = get_token_list(line_read);
+			get_heredoc_list(token_list, &heredoc_list);
+			if (token_list != NULL && !has_pipe_redir_open(token_list))
+				new_redirections(token_list, &g_var.env);
+			lst_str_free(&heredoc_list);
+			megafree(&token_list);
 		}
-		free(str_got);
 	}
+	free(line_read);
 }
 
 int	main(int argc, char **argv, char **env)
@@ -30,11 +79,12 @@ int	main(int argc, char **argv, char **env)
 	(void)argv;
 	g_var.env = NULL;
 	g_var.last_cmd_status = 0;
+	disable_ctrl_c_hotkey();
 	init_ms_env(env, &g_var.env);
 	while (1)
 	{
 		signal_handler_default();
-		processline();
+		processline(env[0] == NULL);
 //		system("lsof -c minishell");
 	}
 	return (0);