+++ src/sandbox/prev_edu.c
@@ -1,0 +1,43 @@
+# include <unistd.h>
+# include <stdlib.h>
+# include <string.h>
+# include <limits.h>
+# include <stdio.h>
+# include <fcntl.h>
+# include <sys/param.h>
+# include <sys/wait.h>
+# include <sys/stat.h>
+# include <readline/readline.h>
+# include <readline/history.h>
+# include <errno.h>
+# include <termios.h>
+
+int main(void)
+{
+	pid_t	pid1;
+	pid_t	pid2;
+	int		fd[2];
+
+	pipe(fd);
+
+	pid1 = fork();
+	if (pid1 == 0)
+	{
+		dup2(fd[1], STDOUT_FILENO);
+		close(fd[0]);
+		close(fd[1]);
+		execlp("ls", "ls", "-la", NULL);
+	}
+	pid2 = fork();
+	if (pid2 == 0)
+	{
+		dup2(fd[0], STDIN_FILENO);
+		close(fd[0]);
+		close(fd[1]);
+		execlp("grep", "grep", "src", NULL);
+	}
+	close(fd[0]);
+	close(fd[1]);
+	waitpid(pid1, NULL, 0);
+	waitpid(pid2, NULL, 0);
+}