+++ src/utils/lst_process_handler.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lst_red_handler.c                                  :+:      :+:    :+:   */
+/*   lst_process_handler.c                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/11/20 11:27:35 by emadriga          #+#    #+#             */
-/*   Updated: 2022/01/16 13:32:53 by emadriga         ###   ########.fr       */
+/*   Created: 2022/01/15 11:27:35 by emadriga          #+#    #+#             */
+/*   Updated: 2022/01/15 16:35:00 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
  * @param list	list
  * @param new	new node to link
 */
-void	lst_red_add_front(t_red **list, t_red *new)
+void	lst_process_add_front(t_pp **list, t_pp *new)
 {
 	new->next = *list;
 	*list = new;
@@ -28,9 +28,9 @@ void	lst_red_add_front(t_red **list, t_red *new)
  * @param list	list
  * @param new	new node to link
 */
-void	lst_red_add_back(t_red **list, t_red *new)
+void	lst_process_add_back(t_pp **list, t_pp *new)
 {
-	t_red	*aux;
+	t_pp	*aux;
 
 	new->next = NULL;
 	if (*list == NULL)
@@ -47,15 +47,18 @@ void	lst_red_add_back(t_red **list, t_red *new)
 /**
  * * Returns malloced node
 */
-t_red	*lst_red_new(void)
+t_pp	*lst_process_new(void)
 {
-	t_red	*output;
+	t_pp	*output;
 
-	output = malloc(sizeof(t_red));
+	output = malloc(sizeof(t_pp));
 	if (!output)
 		return (NULL);
+	output->is_cmd = FALSE;
+	output->argv = NULL;
+	output->pathname = NULL;
+	output->redir = NULL;
 	output->next = NULL;
-//	dprintf(2, "> %p <\n", output);
 	return (output);
 }
 
@@ -63,19 +66,22 @@ t_red	*lst_red_new(void)
  * * Free list
  * @param list	list
 */
-void	lst_red_free(t_red **list)
+void	lst_process_free(t_pp **list)
 {
-	t_red	*next;
-	t_red	*aux;
+	t_pp	*next;
+	t_pp	*aux;
 
 	next = *list;
 	aux = *list;
 	while (next != NULL)
 	{
 		next = next->next;
-		free(aux->path);
-		megafree(&aux->params);
-		megafree(&aux->redirs);
+		if (aux->pathname != NULL)
+			free(aux->pathname);
+		if (aux->argv != NULL)
+			megafree(&aux->argv);
+		if (aux->redir != NULL)
+			lst_redir_free(&aux->redir);
 		free(aux);
 		aux = next;
 	}
@@ -86,28 +92,22 @@ void	lst_red_free(t_red **list)
 /**
  * * Returns malloced node
 */
-void	lst_red_print(t_red *list)
+void	lst_process_print(t_pp *list)
 {
 	int	i;
 
-	dprintf(2, "\nPrinting\n");
+	dprintf(2, "\nPrinting processes\n");
 	while (list != NULL)
 	{
 		i = 0;
-		dprintf(2, "\npid_in %d\tpip_out %d\tpath %s\n Params ", \
-		list->pip_in, list->pip_out, list->path);
-		while (list->params[i])
+		dprintf(2, "\nis_cmd %d\tpathname %s\n argv ", \
+		list->is_cmd, list->pathname);
+		while (list->argv[i])
 		{
-			dprintf(2, "\t param[%d] %s\t", i, list->params[i]);
+			dprintf(2, "\t param[%d] %s\t", i, list->argv[i]);
 			i++;
 		}
-		i = 0;
-		dprintf(2, "\n redirs");
-		while (list->redirs[i])
-		{
-			dprintf(2, "\t redirs[%d] %s\t", i, list->redirs[i]);
-			i++;
-		}
+		lst_redir_print(list->redir);
 		list = list->next;
 	}
 }